+++ coldfront/components/site/templates/common/navbar_admin.html
@@ -15,8 +15,8 @@
     <a id="navbar-allocation-user-requests" class="dropdown-item" href="{% url 'allocation-user-request-list' %}">
       Allocation User Requests
     </a>
-    <a id="navbar-advanced_exporting" class="dropdown-item" href="{% url 'advanced-exporting' %}">
-      Advanced Exporting
+    <a id="navbar-advanced-search" class="dropdown-item" href="{% url 'advanced-search' %}">
+      Advanced Search
     </a>
     <a id="navbar-grant-report" class="dropdown-item" href="{% url 'grant-report' %}">Grant Report</a>
   </div>
