+++ coldfront/config/plugins/advanced_search.py
@@ -1,5 +1,5 @@
 from coldfront.config.base import INSTALLED_APPS
 
 INSTALLED_APPS += [
-    'coldfront.plugins.advanced_exporting'
+    'coldfront.plugins.advanced_search'
 ]
