--- coldfront/plugins/advanced_exporting/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class AdvancedExportingConfig(AppConfig):
-    name = 'coldfront.plugins.advanced_exporting'
