+++ coldfront/plugins/advanced_search/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class AdvancedSearchConfig(AppConfig):
+    name = 'coldfront.plugins.advanced_search'
