+++ coldfront/plugins/advanced_search/templates/advanced_search/advanced_search.html
@@ -20,7 +20,7 @@ Project and Allocation Export List
       </div>
       <div id="collapseOne" class="collapse {{expand_accordion}}" data-parent="#accordion">
         <div class="card-body">
-          <form method="GET" action="{% url 'advanced-exporting' %}" autocomplete="off">
+          <form method="GET" action="{% url 'advanced-search' %}" autocomplete="off">
             {% crispy export_form %}
           </form>
         </div>
