+++ coldfront/plugins/advanced_search/templates/advanced_search/formset.html
