+++ coldfront/plugins/advanced_search/util.py
